Move EventPublishingHandler into the Fahrschein Module as it's not a plain HTTP interface anymore.
